requirements.txt
langchain_core==0.3.49
langchain_ollama==0.3.0
Pillow==11.1.0
